backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
transformers==4.43.3
accelerate==0.33.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
orjson==3.10.7
python-dotenv==1.0.1
